requirements.txt
paho.mqtt
flask
flask_socketio
random
time
firebase_admin